CourseController finished and tested. EnrolmentController and StudentController started

+++ src/main/java/ro/company/controller/CourseController.java
@@ -1,5 +1,6 @@
 package ro.company.controller;
 
+import ro.company.model.Book;
 import ro.company.model.Course;
 import ro.company.repository.CourseRepository;
 
@@ -18,14 +19,38 @@ public class CourseController {
 
     //CRUD
     public boolean add(Course c){
+        if(duplicate(c)){
+            return false;
+        }
         return courseRepository.add(c);
-
+    }
+    public boolean updateCourseName(int id, String newName){
+        Course course = new Course(newName,"Dep1");
+        if(duplicate(course)){
+            return false;
+        }
+        return courseRepository.updateCourseName(id,newName);
+    }
+    public boolean updateDepartmentName(int id, String newName){
+        Course courseD = new Course("C name",newName);
+        if(duplicate(courseD)){
+            return false;
+        } else if(courseRepository.getCourse(id).getDepartment().equals(newName)){
+            return false;
+        }
+        return courseRepository.updateDepartmentName(id,newName);
     }
     public boolean delete(int id){
+        if(containsID(id)==false){
+            return false;
+        }
         return courseRepository.delete(id);
     }
 
     //Utility
+    public Course getCourse(int id){
+        return courseRepository.getCourse(id);
+    }
     public boolean duplicate(Course course){
         if(nbOfDuplicates(course)>=2){
             return true;
@@ -45,5 +70,32 @@ public class CourseController {
     }
     public int lastID(){
         return courseRepository.lastID();
+    }
+    public boolean removeDuplicates(){
+        int count=0;
+        Iterator<Course> it = courseRepository.allCourseList().iterator();
+        while (it.hasNext()){
+            Course c = it.next();
+            if(nbOfDuplicates(c)>=3){
+                count++;
+                courseRepository.delete(c.getId());
+            }
+        }
+        if(count>=1){
+            System.out.println("Total duplicates removed: "+count);
+            return true;
+        }
+        System.out.println("Total duplicates removed: "+count);
+        return false;
+    }
+    public boolean containsID(int id){
+        Iterator<Course> it = courseRepository.allCourseList().iterator();
+        while (it.hasNext()){
+            Course bk = it.next();
+            if(bk.getId()==id){
+                return true;
+            }
+        }
+        return false;
     }
 }
